:bricks: move partners payload collection into misc
