+++ src/payload/collections/misc/PartnersCollection.ts
